py5_resources/py5_module/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py5"
dynamic = ["version"]
description = "Processing for CPython"
readme = "README.md"
license = "LGPL-2.1-only"
requires-python = ">3.9"
authors = [
    { name = "Jim Schmitz", email = "jim@ixora.io" },
]
keywords = [
    "IPython",
    "Jupyter",
    "Processing",
    "Widgets",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)",
    "Programming Language :: Java",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "autopep8>=2.2",
    "jpype1==1.5.0",
    "line_profiler>=4.1",
    "numpy>=1.26",
    "pillow>=10.2",
    "pyobjc>=10.1;sys_platform==\"darwin\"",
    "pywin32>=308;sys_platform==\"win32\"",
    "requests>=2.31",
    "stackprinter>=0.2.11",
]

[project.optional-dependencies]
jupyter = [
    "py5jupyter>=0.2.0a0",
]
extras = [
    "colour>=0.1.5",
    "matplotlib>=3.9",
    "py5jupyter>=0.2.0a0",
    "shapely>=2.0",
    "trimesh>=4.3",
]

[project.scripts]
py5-cmd = "py5_tools.tools.py5cmd:main"
py5-translate-imported2module = "py5_tools.tools.py5translate_imported2module:main"
py5-translate-module2imported = "py5_tools.tools.py5translate_module2imported:main"
py5-translate-processingpy2imported = "py5_tools.tools.py5translate_processingpy2imported:main"
py5-utils = "py5_tools.tools.py5utils:main"
run_sketch = "py5_tools.tools.run_sketch:main"
py5-run-sketch = "py5_tools.tools.run_sketch:main"
py5-live-coding = "py5_tools.tools.live_coding:main"

[project.urls]
"Bug Tracker" = "https://github.com/py5coding/py5generator/issues"
Documentation = "https://py5coding.org/"
Download = "https://pypi.org/project/py5"
Homepage = "https://py5coding.org/"
"Source Code" = "https://github.com/py5coding/py5"

[tool.hatch.version]
path = "py5_tools/constants.py"

[tool.hatch.build]
packages = [
    "/py5",
    "/py5_tools",
]
